Add group information, header, members, and games components
